Added calculate wattage function and implemented it into flowchartresult to feed to frontend. Also made it so the final product doesn't split into seperate machine anymore. Now shared.

+++ SCB.WebApi/Constants/MachineWattage.cs
@@ -4,15 +4,15 @@ using System.Collections.Generic;
 
 public static class MachineWattages
 {
-    public static readonly IReadOnlyDictionary<string, double> Values = new Dictionary<string, double>
+    public static readonly IReadOnlyDictionary<string, int> Values = new Dictionary<string, int>
     {
-        {"Smelter", 4.0},
-        {"Constructor", 4.0},
-        {"Assembler", 15.0},
-        {"Foundry", 16.0},
-        {"Manufacturer", 55.0},
-        {"Refinery", 30.0},
-        {"Blender", 75.0},
-        {"Packager", 10.0}
+        {"Smelter", 4},
+        {"Constructor", 4},
+        {"Assembler", 15},
+        {"Foundry", 16},
+        {"Manufacturer", 55},
+        {"Refinery", 30},
+        {"Blender", 75},
+        {"Packager", 10}
     };
 }
